SSOT/schema alignment, crypto RPC, and build fixes
- Promote missing DB contracts into SSOT and align security notes

- Add missing migrations/tables/RLS and refresh schema snapshot

- Align encryption to pgcrypto via RPC and fix onboarding writes

- Fix TS/build issues; add schema refresh helper and smoke tests

- Upgrade Next.js to patched version to clear npm audit

+++ supabase/migrations/20250210160000_get_trainer_profile_accept_key.sql
@@ -43,7 +43,7 @@ BEGIN
         c.name,
         c.region,
         COALESCE(AVG(r.rating), 0),
-        COUNT(r.id),
+        COUNT(r.id)::INTEGER,
         COALESCE(ARRAY_REMOVE(ARRAY_AGG(DISTINCT ts.age_specialty), NULL), ARRAY[]::age_specialty[]),
         COALESCE(ARRAY_REMOVE(ARRAY_AGG(DISTINCT tbi.behavior_issue), NULL), ARRAY[]::behavior_issue[]),
         COALESCE(ARRAY_REMOVE(ARRAY_AGG(DISTINCT tsvc.service_type), NULL), ARRAY[]::service_type[])
